Fix: uint8_t

+++ cmdline.h
@@ -1,7 +1,3 @@
-//
-// Created by asterwyx on 2020/10/4.
-//
-
 #ifndef RICH_CMDLINE_H
 #define RICH_CMDLINE_H
 
@@ -15,14 +11,14 @@ void tolower(std::string &str);
 int do_preset(std::string cmd);
 int do_roll();
 void do_dump();
-void do_robot(std::uint8_t step, p_player_t player);
-void do_bomb(std::uint8_t step, p_player_t player);
-void do_block(std::uint8_t step, p_player_t player);
-void do_sell(map_t& map, player_t& player, uint8_t map_node_idx);
+void do_robot(std::int step, p_player_t player);
+void do_bomb(std::int step, p_player_t player);
+void do_block(std::int step, p_player_t player);
+void do_sell(map_t& map, player_t& player, int map_node_idx);
 int do_query(player_t& player);
 int do_help();
 void tolower(std::string &str);
 void show_cmd();
-int do_step(std::uint8_t step);
+int do_step(std::int step);
 
 #endif //RICH_CMDLINE_H
